--Empty Message--

+++ blmarket/WellTimedSearch.cpp
@@ -20,26 +20,26 @@ typedef pair<int,int> PII;
 
 template<typename T> int size(const T &a) { return a.size(); }
 
-double go(int N, int A, int B) {
+long double go(int N, int A, int B) {
     if(N == 0) return 0;
-    double px = 1.0 / N;
+    long double px = 1.0 / N;
     if(B == 1) return px;
     if(A == 1) {
-        double p1 = (double)(N / 2) / N;
+        long double p1 = (long double)(N / 2) / N;
         if(N % 2 == 1) {
             return px + 2 * p1 * go(N/2, 1, B-1);
         } else {
-            double p2 = (double)(N-N/2-1) / N;
+            long double p2 = (long double)(N-N/2-1) / N;
             return px + p1 * go(N/2, 1, B-1) + p2 * go(N-1-N/2, 1, B-1);
         }
     }
 
-    double ret = 0;
+    long double ret = 0;
     int md = 0;
     for(int i=0;i<=(N/2);i++) {
-        double p1 = (double)i / N;
-        double p2 = (double)(N-1-i) / N;
-        double tmp = p1 * go(i, A-1, B-1) + p2 * go(N-1-i, A-1, B-1);
+        long double p1 = (long double)i / N;
+        long double p2 = (long double)(N-1-i) / N;
+        long double tmp = p1 * go(i, A-1, B-1) + p2 * go(N-1-i, A-1, B-1);
         if(ret < tmp) { 
             ret = tmp;
             md = i;
@@ -52,7 +52,7 @@ double go(int N, int A, int B) {
 class WellTimedSearch 
 {
 public:
-    double getProbability(int N, int A, int B) 
+    long double getProbability(int N, int A, int B) 
     {
         return go(N, A, B);
     }
@@ -63,11 +63,11 @@ public:
 	void run_test(int Case) { if ((Case == -1) || (Case == 0)) test_case_0(); if ((Case == -1) || (Case == 1)) test_case_1(); if ((Case == -1) || (Case == 2)) test_case_2(); if ((Case == -1) || (Case == 3)) test_case_3(); }
 	private:
 	template <typename T> string print_array(const vector<T> &V) { ostringstream os; os << "{ "; for (typename vector<T>::const_iterator iter = V.begin(); iter != V.end(); ++iter) os << '\"' << *iter << "\","; os << " }"; return os.str(); }
-	void verify_case(int Case, const double &Expected, const double &Received) { cerr << "Test Case #" << Case << "..."; if (Expected == Received) cerr << "PASSED" << endl; else { cerr << "FAILED" << endl; cerr << "\tExpected: \"" << Expected << '\"' << endl; cerr << "\tReceived: \"" << Received << '\"' << endl; } }
-	void test_case_0() { int Arg0 = 3; int Arg1 = 2; int Arg2 = 2; double Arg3 = 0.6666666666666666; verify_case(0, Arg3, getProbability(Arg0, Arg1, Arg2)); }
-	void test_case_1() { int Arg0 = 3; int Arg1 = 3; int Arg2 = 3; double Arg3 = 0.3333333333333333; verify_case(1, Arg3, getProbability(Arg0, Arg1, Arg2)); }
-	void test_case_2() { int Arg0 = 123456; int Arg1 = 1; int Arg2 = 20; double Arg3 = 1.0; verify_case(2, Arg3, getProbability(Arg0, Arg1, Arg2)); }
-	void test_case_3() { int Arg0 = 5; int Arg1 = 3; int Arg2 = 4; double Arg3 = 0.6; verify_case(3, Arg3, getProbability(Arg0, Arg1, Arg2)); }
+	void verify_case(int Case, const long double &Expected, const long double &Received) { cerr << "Test Case #" << Case << "..."; if (Expected == Received) cerr << "PASSED" << endl; else { cerr << "FAILED" << endl; cerr << "\tExpected: \"" << Expected << '\"' << endl; cerr << "\tReceived: \"" << Received << '\"' << endl; } }
+	void test_case_0() { int Arg0 = 3; int Arg1 = 2; int Arg2 = 2; long double Arg3 = 0.6666666666666666; verify_case(0, Arg3, getProbability(Arg0, Arg1, Arg2)); }
+	void test_case_1() { int Arg0 = 3; int Arg1 = 3; int Arg2 = 3; long double Arg3 = 0.3333333333333333; verify_case(1, Arg3, getProbability(Arg0, Arg1, Arg2)); }
+	void test_case_2() { int Arg0 = 123456; int Arg1 = 1; int Arg2 = 20; long double Arg3 = 1.0; verify_case(2, Arg3, getProbability(Arg0, Arg1, Arg2)); }
+	void test_case_3() { int Arg0 = 5; int Arg1 = 3; int Arg2 = 4; long double Arg3 = 0.6; verify_case(3, Arg3, getProbability(Arg0, Arg1, Arg2)); }
 
 // END CUT HERE
